+++ src/db/src/DB.php
@@ -80,6 +80,16 @@ class DB
     }
 
     /**
+     * Make a new connection with the pool name.
+     */
+    public static function connection(string $poolName): self
+    {
+        return make(static::class, [
+            'poolName' => $poolName,
+        ]);
+    }
+
+    /**
      * Define the commands that needs same connection to execute.
      * When these commands executed, the connection will storage to coroutine context.
      */