+++ src/pool/src/SimplePool/PoolFactory.php
@@ -70,4 +70,9 @@ class PoolFactory
     {
         return $this->configs[$name];
     }
+
+    public function getPoolNames(): array
+    {
+        return array_keys($this->pools);
+    }
 }